Update README/app to allow for more demo pages
In the midst of trying to get the Makefile working/etc., I'd somewhat
lost the thread of why I was making this in the first place; while the
learning/etc. was a bonus, the original target goal was actually to demo
how easy it is to get PostgreSQL data into a JSON string for feeding out
from a web API.

The README has been updated to reflect this, and the app code
restructured so I can easily include additional examples.