Add "raw table" example
This example shows the simple contents of a table in a Google Chart
"Table" visualization. It is the most simple transform on the SQL side
that I can imagine, and the JS is (aside from the call to map())
additionally relatively simple as well.

It also updates the README to include a reference to the new table full
of auto-generated fake data that we're creating as a part of this
modification.