Add view definition on 'ps_checker' page
Because this view is particularly complicated, and I originally didn't
want to show off an overly-complex query that might scare people off, I
created a view. I'd still like to show a view, as a way to "simplify"
some really complex operations, but I'd also like to show the view
itself.

And, rather than copy and paste it, I figured it was easier just to
scrape the file that creates it, and show it as an additional section on
only the '/ps_checker' page.